Moved tab-container to correct subfolder.
Rearranged header
